Fix update.ts

+++ src/update.ts
@@ -9,25 +9,33 @@ const updatedPrefixes: prefixID[] = ["Snowy", "Endangered", "Emburdening"];
 const iconDirectory = `./../ccicons/`
 const allFiles = readDirectoryRecursively(iconDirectory);
 
+console.log(allFiles.filter(file => file.includes('prefix') && !file.includes('.')))
+
 const updatedCubeFiles = [...allFiles].filter(file => updatedCubes.find(cubeID => file.includes(`${cubeID}.png`) || file.includes(`${cubeID}.gif`) ) !== undefined)
 
-const updatedPrefixFiles = [...allFiles].filter(file => updatedPrefixes.find(prefixID => !file.includes('.') && !file.includes('\\') && file.includes(`prefix`) && file.includes(prefixID.toLowerCase())) !== undefined)
+const updatedPrefixFiles = [...allFiles].filter(file => updatedPrefixes.find(prefixID => !file.includes('.') && file.includes(`prefix`) && file.includes(prefixID.toLowerCase())) !== undefined)
 
 const updatedPrefixCacheDirs = [...allFiles].filter(file => file.startsWith('prefixcache') && !file.endsWith('.png') && !!updatedPrefixes.find(prefixID => file.includes(prefixID.toLowerCase())))
 
 updatedCubeFiles.forEach(filePath => {
-    fs.unlinkSync(filePath);
-    console.log(`Deleted Image: ${filePath}`);
+    if (fs.existsSync(filePath)) {
+        fs.unlinkSync(filePath);
+        console.log(`Deleted Image: ${filePath}`);
+    }
 });
 
 updatedPrefixFiles.forEach(filePath => {
-    fs.rmSync(filePath, { recursive: true, force: true });
-    console.log(`Deleted Prefix Directory: ${filePath}`);
+    if (fs.existsSync(filePath)) {
+        fs.rmSync(filePath, { recursive: true, force: true });
+        console.log(`Deleted Prefix Directory: ${filePath}`);
+    }
 });
 
 updatedPrefixCacheDirs.forEach(filePath => {
-    fs.rmSync(filePath, { recursive: true, force: true });
-    console.log(`Deleted Prefix Cache Directory: ${filePath}`);
+    if (fs.existsSync(filePath)) {
+        fs.rmSync(filePath, { recursive: true, force: true });
+        console.log(`Deleted Prefix Cache Directory: ${filePath}`);
+    }
 });
 
 console.log(`Update Complete.
